Add collision handling system. See detailed desc.
Brief description of system: creates a virtual rectangle around the
Player sprite, with 5 pixels in between each edge of the sprite and the
rectangle.  For each corner of the rectangle, the tile location is
determined and from it.  The entities near that tile are determined, and
if one is a barrier entity, the player's speed in the relevant direction
is adjusted accordingly.  Handling of the collision with barrier
entities (such as door unlocking and combat initialization) can, and
will, be added.